The modifications to the fast-api-server that allowed the Swagger demo against the Django Oauth Server on Nov 13, 2024.
server.py add middleware that helps authlib store session state.

nlip.py adds a security protection on the /nlip endpoint that verifies that there is a current user
and that their token is valid.  It checks the validity of the token by querying against the /userinfo
endpoint of the authentication server.